Add badge template that can be used for scout_group badges.

+++ data_badge.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-"""Badge class används för märken och cerifikat och deras olika delar."""
 import logging
 from collections import namedtuple
 
@@ -7,19 +6,18 @@ from google.appengine.api import memcache, users
 from google.appengine.ext import ndb
 
 
-from data import ScoutGroup, Troop, Person, PropertyWriteTracker
-
-ADMIN_OFFSET = 100  # Offset for administrative badge parts
+from data import ScoutGroup, Troop, Person
 
 
 class BadgeImage(ndb.Model):
-    "Image/thumbnail for a badge. Stored as data URL."
-    image = ndb.TextProperty(required=True)  # TODO. Use BlobProperty to save space (get rid of unicode)
+    """Image/thumbnail for a badge. Stored as data URL."""
+    image = ndb.TextProperty(required=True)
     name = ndb.StringProperty()  # Name of badge
+    isTemplate = ndb.BooleanProperty(required=True)  # Owned by template. Cannot be removed by others
 
 
 class Badge(ndb.Model):
-    "Badge definition for a scout group (scoutkår). The required parts are separate as BadgePart."
+    """Badge definition for a scout group (scoutkår). The required parts are separate as BadgePart."""
     name = ndb.StringProperty(required=True)
     scoutgroup = ndb.KeyProperty(kind=ScoutGroup, required=True)
     description = ndb.StringProperty(required=True)
@@ -32,7 +30,7 @@ class Badge(ndb.Model):
     image_key = ndb.KeyProperty(kind=BadgeImage)
 
     @staticmethod
-    def create(name, scoutgroup_key, description, parts_scout, parts_admin, img_data):
+    def create(name, scoutgroup_key, description, parts_scout, parts_admin, img_data=None, img_key=''):
         if name == "" or name is None:
             logging.error("No name set for new badge")
             return
@@ -48,15 +46,17 @@ class Badge(ndb.Model):
                       parts_admin_short=parts_admin_short,
                       parts_admin_long=parts_admin_long)
         if img_data:
-            image = BadgeImage(image=img_data, name=name)
+            image = BadgeImage(image=img_data, name=name, isTemplate=False)
             badge.image_key = image.put()
+        elif img_key:
+            badge.image_key = img_key
         badge.put()
 
     def update(self, name, description, parts_scout, parts_admin, img_data):
         """Update badge parts for badge. Separate scout series from admin."""
         changed = False
         if name == "" or name is None:
-            logging.info("No name set for badge update")
+            logging.warning("No name set for badge update")
             return
         if name != self.name:
             self.name = name
@@ -98,15 +98,17 @@ class Badge(ndb.Model):
             changed = True
 
         if img_data:
-            logging.info("New image data")
+            # logging.info("New image data")
             if self.image_key is not None:
-                self.image_key.delete()
-            image = BadgeImage(image=img_data, name=name)
+                image = self.image_key.get()
+                if not image.isTemplate:
+                    self.image_key.delete()
+            image = BadgeImage(image=img_data, name=name, isTemplate=False)
             self.image_key = image.put()
             changed = True
 
         if changed:
-            logging.info("Badge %s updated" % self.name)
+            # logging.info("Badge %s updated" % self.name)
             self.put()
 
     @staticmethod
@@ -116,42 +118,55 @@ class Badge(ndb.Model):
             badges = Badge.query(Badge.scoutgroup == scoutgroup_key).order(Badge.name).fetch()
         return badges
 
-    def update_for_person(self, person_key, idx_list, examiner_name):
+    def update_for_person(self, person_key, scout_idx_list, admin_idx_list, examiner_name):
         "Add new idx and create BadgeCompleted when all parts done."
-        person = person_key.get()
-        logging.info("Badge %s parts %s for %s" % (self.name, idx_list, person.getname()))
+        # logging.info("Badge %s parts %s for %s" % (self.name, idx_list, person.getname()))
         badge_key = self.key
         parts_done = BadgePartDone.parts_done(person_key, badge_key)
-        prev_idx = [pd.idx for pd in parts_done]
+        prev_scout_idx = [pd.idx for pd in parts_done if pd.is_scout_part]
+        prev_admin_idx = [pd.idx for pd in parts_done if not pd.is_scout_part]
+        nr_prev_parts = len(prev_scout_idx) + len(prev_admin_idx)
+        nr_all_parts = len(self.parts_scout_short) + len(self.parts_admin_short)
+
         nr_idx_added = 0
-        all_idx = range(len(self.parts_scout_short))
-        for idx in range(ADMIN_OFFSET, ADMIN_OFFSET+len(self.parts_admin_short)):
-            all_idx.append(idx)
-        for idx in idx_list:
-            if idx in prev_idx:
-                logging.warn("Trying to set idx %d again", idx)
-            if idx not in all_idx:
-                logging.warn("Trying to set bad idx %d" % idx)
-            BadgePartDone.create(person_key, badge_key, idx, examiner_name)
+        for idx in scout_idx_list:
+            if idx in prev_scout_idx:
+                logging.warn("Trying to set scout idx %d again", idx)
+                continue
+            if idx >= len(self.parts_scout_short):
+                logging.warn("Trying to set scout idx %d beyond top", idx)
+                continue
+            BadgePartDone.create(person_key, badge_key, idx, True, examiner_name)
             nr_idx_added += 1
-        if nr_idx_added > 0 and nr_idx_added + len(prev_idx) == len(all_idx):
+        for idx in admin_idx_list:
+            if idx in prev_admin_idx:
+                logging.warn("Trying to set admin idx %d again", idx)
+                continue
+            if idx >= len(self.parts_admin_short):
+                logging.warn("Trying to set admin idx %d beyond top", idx)
+                continue
+            BadgePartDone.create(person_key, badge_key, idx, False, examiner_name)
+            nr_idx_added += 1
+        if nr_idx_added > 0 and nr_idx_added + nr_prev_parts == nr_all_parts:
             bc = BadgeCompleted(badge_key=badge_key, person_key=person_key, examiner=examiner_name)
             bc.put()
-            logging.info("Badge %s completed by %s" % (self.name, examiner_name))
+            # logging.info("Badge %s completed by %s" % (self.name, examiner_name))
 
 
 class BadgePartDone(ndb.Model):
     "Part that has been done including date and who registered in Skojjt."
     badge_key = ndb.KeyProperty(kind=Badge, required=True)
     person_key = ndb.KeyProperty(kind=Person, required=True)
-    idx = ndb.IntegerProperty(required=True)  # idx for BadgePart
+    idx = ndb.IntegerProperty(required=True)  # idx for BadgePart (zero-based)
+    is_scout_part = ndb.BooleanProperty(required=True)
     date = ndb.DateTimeProperty(auto_now_add=True)
     examiner_name = ndb.StringProperty(required=True)
 
     @staticmethod
-    def create(person_key, badge_key, badge_part_idx, examiner_name):
+    def create(person_key, badge_key, badge_part_idx, is_scout_part, examiner_name):
         bpd = BadgePartDone(person_key=person_key, badge_key=badge_key,
-                            idx=badge_part_idx, examiner_name=examiner_name)
+                            idx=badge_part_idx, is_scout_part=is_scout_part,
+                            examiner_name=examiner_name)
         bpd.put()
 
     @staticmethod
@@ -187,7 +202,7 @@ class TroopBadge(ndb.Model):
         old_troop_badges = [Badge.get_by_id(tp.badge_key.id()) for tp in old_troop_badge_ids]
         old_troop_badges = sorted(old_troop_badges, key=lambda x: x.name)
         nr_old_troop_badges = len(old_troop_badges)
-        logging.info("New are %d, old were %d" % (len(name_list), nr_old_troop_badges))
+        # logging.info("New are %d, old were %d" % (len(name_list), nr_old_troop_badges))
         # First find keys to remove
         to_remove = []
         for old in old_troop_badges:
@@ -208,7 +223,7 @@ class TroopBadge(ndb.Model):
                 really_new.append(name)
         if len(really_new) == 0:
             return
-        logging.info("Really new are %d" % len(really_new))
+        # logging.info("Really new are %d" % len(really_new))
         allbadges = Badge.get_badges(troop.scoutgroup)
         idx = nr_old_troop_badges
         for badge in allbadges:
@@ -216,3 +231,98 @@ class TroopBadge(ndb.Model):
                 tb = TroopBadge(troop_key=troop.key, badge_key=badge.key)
                 tb.put()
                 idx += 1
+
+
+class BadgeTemplate(ndb.Model):
+    """Badge template/blue print to start make a Badge for ascout group (scoutkår)."""
+    name = ndb.StringProperty(required=True)
+    description = ndb.StringProperty(required=True)
+    # Short and long descriptions of scout parts (idx = 0, 1, 2, ...99)
+    parts_scout_short = ndb.StringProperty(repeated=True)
+    parts_scout_long = ndb.StringProperty(repeated=True)
+    # Short and long descriptions of admin parts (idx=100, 101, 101, ...)
+    parts_admin_short = ndb.StringProperty(repeated=True)
+    parts_admin_long = ndb.StringProperty(repeated=True)
+    image_key = ndb.KeyProperty(kind=BadgeImage)
+
+    @staticmethod
+    def get_templates():
+        return BadgeTemplate.query().order(BadgeTemplate.name).fetch()
+
+    @staticmethod
+    def create(name, description, parts_scout, parts_admin, img_data=None):
+        if name == "" or name is None:
+            logging.error("No name set for new badge")
+            return
+        # logging.info(parts_admin)
+        parts_scout_short = [p[0] for p in parts_scout]
+        parts_scout_long = [p[1] for p in parts_scout]
+        parts_admin_short = [p[0] for p in parts_admin]
+        parts_admin_long = [p[1] for p in parts_admin]
+        template = BadgeTemplate(name=name,
+                                 description=description,
+                                 parts_scout_short=parts_scout_short,
+                                 parts_scout_long=parts_scout_long,
+                                 parts_admin_short=parts_admin_short,
+                                 parts_admin_long=parts_admin_long)
+        if img_data:
+            image = BadgeImage(image=img_data, name=name, isTemplate=True)
+            template.image_key = image.put()
+        template.put()
+
+    def update(self, name, description, parts_scout, parts_admin, img_data):
+        """Update badge parts for badge. Separate scout series from admin."""
+        changed = False
+        if name == "" or name is None:
+            logging.warning("No name set for badge update")
+            return
+        if name != self.name:
+            self.name = name
+            changed = True
+
+        if description != self.description:
+            self.description = description
+            changed = True
+
+        if len(parts_scout) < len(self.parts_scout_short):
+            # Remove parts not supported yet. Should also remove BadgePartDone
+            logging.error("Removing scout parts in update, not yet supported")
+            return
+
+        if len(parts_admin) < len(self.parts_admin_short):
+            # Remove parts not supported yet. Should also remove BadgePartDone
+            logging.error("Removing admin parts in update, not yet supported")
+            return
+
+        parts_scout_short = [p[0] for p in parts_scout]
+        parts_scout_long = [p[1] for p in parts_scout]
+        parts_admin_short = [p[0] for p in parts_admin]
+        parts_admin_long = [p[1] for p in parts_admin]
+
+        if parts_scout_short != self.parts_scout_short:
+            self.parts_scout_short = parts_scout_short
+            changed = True
+
+        if parts_scout_long != self.parts_scout_long:
+            self.parts_scout_long = parts_scout_long
+            changed = True
+
+        if parts_admin_short != self.parts_admin_short:
+            self.parts_admin_short = parts_admin_short
+            changed = True
+
+        if parts_admin_long != self.parts_admin_long:
+            self.parts_admin_long = parts_admin_long
+            changed = True
+
+        if img_data:
+            # logging.info("New image data")
+            if self.image_key is not None:
+                self.image_key.delete()
+            image = BadgeImage(image=img_data, name=name, isTemplate=True)
+            self.image_key = image.put()
+            changed = True
+
+        if changed:
+            # logging.info("Badge %s updated" % self.name)
+            self.put()
